+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Swift
   * Make `EventMetricType`, `ObjectMetricType`, `URLMetricType` and `Ping` `Sendable`  ([#3255](https://github.com/mozilla/glean/pull/3255))
+  * Glean for iOS is now being built with Xcode 16.4 ([#3270](https://github.com/mozilla/glean/pull/3270))
 
 # v65.2.1 (2025-09-26)